fix(trading): move market data hook inside component
