+++ libs/market-list/src/lib/market-data-provider.ts
@@ -98,13 +98,3 @@ export const useStaticMarketData = (marketId?: string, skip?: boolean) => {
   });
   return data;
 };
-
-export const useMarketData = (marketId?: string, skip?: boolean) => {
-  const variables = useMemo(() => ({ marketId }), [marketId]);
-  const { data } = useDataProvider({
-    dataProvider: marketDataProvider,
-    variables,
-    skip: skip || !marketId,
-  });
-  return data;
-};
